windows build fix

+++ lib-SLProject/cuda/SLOptixRaytracerShading.cu
@@ -96,7 +96,7 @@ extern "C" __global__ void __closesthit__radiance() {
 
         // Shading like whitted
         color = color +
-                (material.specular_color * pow( max(dot(R, V), 0.0), material.shininess ) * (1.0f - occlusion) // specular
+                (material.specular_color * powf( max(dot(R, V), 0.0), material.shininess ) * (1.0f - occlusion) // specular
                 + material.diffuse_color * max(nDl, 0.0f) * (1.0f - occlusion)) // diffuse
                 * params.lights[i].color; // multiply with light color
     }
